+++ packages/dart_edge_core/lib/src/http/response_spec.dart
@@ -33,6 +33,12 @@ final class ResponseSpec {
   const ResponseSpec.html({int status = 200})
     : this._(status: status, contentType: 'text/html; charset=utf-8');
 
+  /// Creates a binary response specification.
+  const ResponseSpec.binary({
+    int status = 200,
+    String contentType = 'application/octet-stream',
+  }) : this._(status: status, contentType: contentType);
+
   /// Creates a server-sent events response specification.
   const ResponseSpec.sse({int status = 200})
     : this._(status: status, contentType: 'text/event-stream; charset=utf-8');