+++ packages/dart_edge_http_client/lib/src/dart_edge_http_client_transport.dart
@@ -51,7 +51,7 @@ final class DartEdgeHttpClientTransport implements DartEdgeClientTransport {
       status: response.statusCode,
       contentType: response.headers['content-type'] ?? '',
       headers: response.headers,
-      body: response.body,
+      bodyBytes: response.bodyBytes,
     );
   }